Let to the user handle the qrcode
